requirements.txt
# Python packages
# -----------------------------------------------------------------------------
pytz  # https://github.com/stub42/pytz
python-slugify==6.1.2
ipython==7.34.0
ptyprocess==0.7.0

# Django
# -----------------------------------------------------------------------------
Django  # https://www.djangoproject.com/
django-debug-toolbar==3.7.0
django-extensions==3.2.1
Werkzeug==2.2.2
django-environ==0.9.0
# Django Vite
# -----------------------------------------------------------------------------
django-vite==2.0.2

# Django Rest framework
# -----------------------------------------------------------------------------
djangorestframework==3.14.0
django-cors-headers==3.13.0

# Django caching
# -----------------------------------------------------------------------------

# Gunicorn
# -----------------------------------------------------------------------------
gunicorn==20.1.0
uvicorn[standard]

# Util module for custom readines check for the pod
# -----------------------------------------------------------------------------
psutil==5.9.2